# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Vet Appointments API",
    "description": "First version of API to manage vet appointments."
  },
  "host": "localhost:3001",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/veterinarians/": {
      "get": {
        "tags": [
          "Veterinarians"
        ],
        "description": "Get all veterinarians",
        "parameters": [],
        "responses": {
          "200": {
            "description": "all veterinarians.",
            "schema": {
              "$ref": "#/definitions/Veterinarians"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Veterinarians"
        ],
        "description": "add a vet",
        "parameters": [
          {
            "name": "newVet",
            "in": "body",
            "description": "vet information",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Veterinarian"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "new vet.",
            "schema": {
              "$ref": "#/definitions/Veterinarian"
            }
          },
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/veterinarians/{vetId}": {
      "get": {
        "tags": [
          "Veterinarians"
        ],
        "description": "Get veterinarian by Id",
        "parameters": [
          {
            "name": "vetId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Veterinarian by Id.",
            "schema": {
              "$ref": "#/definitions/Veterinarian"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Veterinarians"
        ],
        "description": "Delete a vet",
        "parameters": [
          {
            "name": "vetId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Id of vet"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "put": {
        "tags": [
          "Veterinarians"
        ],
        "description": "Update a vet",
        "parameters": [
          {
            "name": "vetId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "newVet",
            "in": "body",
            "description": "vet information",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Veterinarian"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/veterinarians/{vetId}/slots": {
      "get": {
        "tags": [
          "Veterinarians"
        ],
        "description": "Get veterinarian",
        "parameters": [
          {
            "name": "vetId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Slots by Veterinarian Id.",
            "schema": {
              "$ref": "#/definitions/Slots"
            }
          }
        }
      }
    },
    "/species/": {
      "get": {
        "tags": [
          "Species"
        ],
        "description": "Get all Species",
        "parameters": [],
        "responses": {
          "200": {
            "description": "All Species.",
            "schema": {
              "$ref": "#/definitions/Species"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Species"
        ],
        "description": "add a Species",
        "parameters": [
          {
            "name": "newSpecies",
            "in": "body",
            "description": "Species information",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Species"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "new Species.",
            "schema": {
              "$ref": "#/definitions/Species"
            }
          },
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/species/{speciesId}": {
      "get": {
        "tags": [
          "Species"
        ],
        "description": "Get Species by Id",
        "parameters": [
          {
            "name": "speciesId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Species by Id.",
            "schema": {
              "$ref": "#/definitions/Species"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Species"
        ],
        "description": "Delete a Species",
        "parameters": [
          {
            "name": "speciesId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Id of Species"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "put": {
        "tags": [
          "Species"
        ],
        "description": "Update a Species",
        "parameters": [
          {
            "name": "speciesId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "newSpecies",
            "in": "body",
            "description": "Species information",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Species"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/species/{speciesId}/patients": {
      "get": {
        "tags": [
          "Species"
        ],
        "description": "Get patients by species id",
        "parameters": [
          {
            "name": "speciesId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Patients by Species Id.",
            "schema": {
              "$ref": "#/definitions/Species"
            }
          }
        }
      }
    },
    "/patients/": {
      "get": {
        "tags": [
          "Patients"
        ],
        "description": "Get all patients",
        "parameters": [],
        "responses": {
          "200": {
            "description": "all patients.",
            "schema": {
              "$ref": "#/definitions/Patients"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Patients"
        ],
        "description": "add a patient",
        "parameters": [
          {
            "name": "newPatient",
            "in": "body",
            "description": "patient information",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Patient"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "new patient.",
            "schema": {
              "$ref": "#/definitions/Patient"
            }
          },
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/patients/{patientId}": {
      "get": {
        "tags": [
          "Patients"
        ],
        "description": "Get patients by Id",
        "parameters": [
          {
            "name": "patientId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Patient by Id.",
            "schema": {
              "$ref": "#/definitions/Patient"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Patients"
        ],
        "description": "Delete a patient",
        "parameters": [
          {
            "name": "patientId",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Id of patient"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "put": {
        "tags": [
          "Patients"
        ],
        "description": "Update a patient",
        "parameters": [
          {
            "name": "patientId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "newPatient",
            "in": "body",
            "description": "patient information",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Patient"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/patients/{patientId}/appointments": {
      "get": {
        "tags": [
          "Patients"
        ],
        "description": "Get appointments by patient",
        "parameters": [
          {
            "name": "patientId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Appointments by patient.",
            "schema": {
              "$ref": "#/definitions/Appointments"
            }
          }
        }
      }
    },
...
    "/slots/{slotId}/appointments": {
      "get": {
        "tags": [
          "Slots"
        ],
        "description": "Get appointments by slot",
        "parameters": [
          {
            "name": "slotId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Appointments by Slot Id.",
            "schema": {
              "$ref": "#/definitions/Appointment"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "Veterinarian": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "John Snow"
        },
        "regNumber": {
          "type": "string",
          "example": "11112"
        }
      }
    },
    "Veterinarians": {
      "type": "array",
      "items": {
        "type": "object",
        "properties": {
          "vetId": {
            "type": "string",
            "example": "a2s44"
          },
          "name": {
            "type": "string",
            "example": "John Snow"
          },
          "regNumber": {
            "type": "string",
            "example": "11112"
          }
        }
      }
    },
    "Species": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Dinosaur"
        }
      }
    },
...
    "Slots": {
      "type": "array",
      "items": {
        "type": "object",
        "properties": {
          "slotId": {
            "type": "string",
            "example": "er332"
          },
          "vetId": {
            "type": "string",
            "example": "23123"
          },
          "time": {
            "type": "number",
            "example": 9
          }
        }
      }
    },
...
  }
}